ft: move SERVER_API_VERSION to client constants
