+++ tests/server/services/test_telemetry.py
@@ -30,7 +30,7 @@ def error_sens_o_matic_mock():
 
 
 async def test_sens_o_matic_called_correctly(sens_o_matic_mock):
-    from prefect.server.api.server import SERVER_API_VERSION
+    from prefect.client.constants import SERVER_API_VERSION
 
     telemetry = Telemetry(handle_signals=False)
     await telemetry.start(loops=1)
